Refactoring train_udf.py and test_train_udf.py - Extract Model and ColumnPreprocessor creation out of train_udf.py - Extract langauge container fixture from train_udf.py - Refactor parts of test_train_udf.py into functions
