+++ tests/conftest.py
@@ -1,0 +1,3 @@
+pytest_plugins = [
+    "tests.integration_tests.fixtures.build_language_container_fixture",
+]
